Omit VMIProperties attributes when not set
This commit fixes a bug discovered during the Stratosphere Demo
presentation on Feb 2, 2023 when an unknown bug occurred on Azure
Server side returning the `VMIProperties` without all attributes set.

This led our request to send a value with `None` (null) which couldn't
be parsed by the API and raised errors.

To avoid this situation all boolean attributes on `VMIProperties` are
now configured to be omitted in the request when not set, thus using the
Microsoft defaults for it.