Include Azure Product Ingestion API Schemas
This commit add the Azure Product Ingestion API schemas to the local
repository in order to allow the developers to check for changes on it
along the time.

It also includes a simple script to download and store the schemas from
Microsoft which is called by the CI for each run in order to identify
whether new changes are made, meaning that probably we'll have to update
our library.