Include missing fields and entities in modules
This commit includes optional attributes for the existing models,
based in Microsoft schemas.

As the missing fields are not mandatory they're all marked to be
omitted when not set.

The following models have been updated with optional fields:
    - TestDrive
    - CustomerLeads
    - ProductSummary
    - Listing
    - ListingAsset
    - PlanSummary
    - PlanListing
    - PriceAndAvailabilityPlan
    - DiskVersion
    - ProductSubmission

Also new models were created to represent some of the optional fields,
being them:

  Related with `ProductSummary`, `ProductSubmission`, `PlanSummary`:
    - DeprecationAlternative
    - DeprecationSchedule

  Related with `CustomerLeads`
    - LeadConfiguration
    - BlobLeadConfiguration
    - DynamicsLeadConfiguration
    - EmailLeadConfiguration
    - HttpsEndpointLeadConfiguration
    - MarketoLeadConfiguration
    - SalesforceLeadConfiguration
    - TableLeadConfiguration

  Related with PriceAndAvailabilityPlan
    - SoftwareReservation
    - SoftwareTrial

  Related with VMImageSource
    - DataDisk

The documentation is also updated to reflect the new changes.