Allow AzureService to overwrite a plan with VM img
This commit allows the AzureService to process the `overwrite` key from
the incoming metadata in order to erase all existing VM images in a
particular plan and only setting the new one.

If overwrite is set to `False` it will behave exactly as it was before:
just append the image if the DiskVersion.version number is not the same.