Ensure the DiskVersion is appended when needed
In the previous versions of AzureService a given DiskVersion would be
replaced instead of appeneded when a different number and SAS URI are
given.

The intended behavior is to "append" the images when overwrite == False
and not replace it.

This commit fixes this behavior by ensuring the inner SAS get changed
only when the DiskVersion matches the expected version.

The fix consists of:

1 - Just process a change when the SAS URI is not set. For this the
  behavior of `is_sas_uri` was changed to scan all disk versions to
ensure the SAS is not there.

2 - Change the `_seek_disk_version` to just return the DiskVersion when
  it matches the version number from metadata. With this we'll not
overwrite the content of a wrong disk version.

3 - Update the tests