Split the AzureService.publish even more
Split the method `publish` even more with the new private method
`_set_new_sas_disk_version` which is now responsible to set a new SAS
URI in the Disk Version when the version number matches the metadata.

The goal of this commit is to simplify even more the workflow
legibility.