Azure: always overwrite images on draft
This commit changes the Azure workflow to always set a new disk_version
whenever the state of the plan is draft, since it's impossible to have
two or more images pending publish in the same plan.

With this commit the option `overwrite` will be automatically `True` and
not required at all from StArMap (it won't have any effect).