EXDSP-2305: Do not filter out deprecated images
A new bug was discovered on EXDSP-2305 when sending a `configure`
request to add a new image and filtering out deprecated images.

It turns out that for some reason Microsoft now requires us to send ALL
published images, even the deprecated ones.

Since the deprecated schema doesn't require `date_offset` to be sent we
marked it as optional and not visible when not set.

This commit fully fixes the bug described on EXDSP-2305.