Sanitize scheduled deprecation images on Azure
This commit implements a filter function to sanitize the scheduled
deprecation images on Azure since the Microsoft API is misbehaving with
this case.

We can not delete the scheduled deprecation image nor include it as is.

We need to get rid of it's lifecycle state and deprecation schedule.

Refers to EXDSP-2380