No longer filter out deprecated images
This commit deimplements the filter for deprecated images, as Microsoft
has updated their server and now requires passing all deprecatated and
scheduled to be deprecated images as they were.

By appliying the filter implemented on cloudpub!52 now causes the server
to interpret that as "cancel the deprecation", which may cause the
following error:

```
Image ... cannot be restored because it was deprecated before it
was published to live
```

I've tested the results using Postman and the patched version to ensure
this is the expected behavior from Azure after their change.

Refers to EXDSP-2544