Add default "reason" for DeprecationSchedule
This commit is a "workaround" on `DeprecationSchedule.reason` to always
have a default value for it.

Although the field `reason` is a mandatory one, for some reason
(probably manual changes), the offer
`2d541119-30c2-4a31-ada6-4e9b38d86697` from AZURE-EMEA has the reason
set as `null` which was breaking the request to `configure` causing the
error:

```
Schema validation failed for schema https://product-ingestion.azureedge.net/schema/virtual-machine-plan-technical-configuration/2022-03-01-preview3: #/vmImageVersions/3/deprecationSchedule/reason: Value is \u0022null\u0022 but should be \u0022string\u0022;#/vmImageVersions/3/deprecationSchedule/reason: Expected value to match one of the values specified by the enum;
```

With this commit we ensure the attribute has a value even if it's messed
by manual changes.

Refers to EXDSP-2544