Azure: Check first for draft instead live
This commit changes the `get_product` order to retrieve the information
using the following criteria:

```
preview -> draft -> live
```

Instead of the previous

```
preview -> live -> draft
```

This is necessary since according to https://pub.devel.redhat.com/pub/task/713800/log/stdout.log
the `ConfigureRequest` to add the disk version was sent twice per
region as the product was being retrieved with "live" which didn't
contain the changes performed in "pre-push", present in "draft" only.

With this we hope to avoid wasting resources by checking for changes
before asking again to perform the ones.

It also may fix the issue described on JBEAP-26231, which states about a
duplicated submission to "preview". However, since there's no evidence
in the logs that it was caused because of this configuration (as there's
no repeated "preview assigment" for each region) it MAY solve but it's
not a ceirtanty. Most probably the issue was caused either by a
connection retry or some manual interference in the process.