Azure: Allow custom target for get_product
This commit changes the `AzureService` methods `get_product` and
`get_product_by_name` to have an optional parameter called
`first_target` which is used to modify at runtime the target priority
when retrieving the product data from Azure.

In order to be retrocompatible the default parameter for `first_target`
is `preview`.

With this change it will be possible for the library to adjust the
target retrieval when a corner use case requires such operation, like
the "stage to preview" feature.