Fix stage to preview workflow
This fix allows a product marked to be published to `preview` during its
non live push (`--keep-draft` / `--nochannel`) to be in fact published
when the offer is already in preview.

The desired behavior for this situation is:

1 - Add the image to the plan
2 - Always submit to preview

However, if the offer was already in `preview` state it wasn't going to
publish again to `preview` due to idepontency.

To fix that we just need to make sure the first target for `get_product`
when `stage preview` is set would be `draft`, then it will ignore the
already exissting `preview` state and do the publishing end to end.