Azure: Include SV in SAS comparison exclusion
This commit includes the `signedVersion` (`sv`) parameter from SAS URI
into the comparison exclusion in order to ensure that 2 SAS URIs with
the same base URL and different values for `sv` are considered
equivalent.

It fixes the "multiple SAS for the same image" issue found while
pushing `jboss-eap` images.

Refers to EXDSP-2538