Allow defining fixed values for models attributes
This commit changes the `AttrsJSONDecodeMixin` to allow defining a
constant for any field by using the metadata keyword `const`.

The idea is that any value defined on `const` will override the received
value so it will be possible to "pin" values whenever we like instead of
relying on what comes from the server.

Refers to SPSTRAT-144