Allow publishing to draft plans in Azure
This commit creates a new method named `get_product_plan_by_name` which
seeks the desired plan name in the product for all available Azure
Targets: `draft`, `live`, `preview`.

It also changes the `publish` method to use this new method to prevent a
bug which a product is found but not the desired plan because it is in
the `draft` target.

Refers to SPSTRAT-296