Azure: Pin ProductSubmission schema
This commit pins the `ProductSubmission` schema to the version
`2022-03-01-preview2`.

It prevents the publishing to fail with the error:

```
Schema validation failed for schema https://schema.mp.microsoft.com/schema/submission/2022-03-01-preview7: Required properties [\u0022sources\u0022] were not present #
```

Introduced by the version `2022-03-01-preview7` which now requires a new
property `Source` defined below:

```
"source": {
  "type": "object",
  "properties": {
    "module": {
      "type": "string",
      "enum": [
        "ageRating",
        "availability",
        "listing",
        "package",
        "plans",
        "property",
        "resell",
        "supplementalContent",
        "xbox",
        "xboxMetadata"
      ]
    },
    "branch": {
      "type": "string",
      "default": "main",
      "minLength": 1,
      "maxLength": 255
    }
    },
  "additionalProperties": false,
  "required": [
    "module"
  ]
}
```

While it's not yet clear what we should send in this case from the Enum,
as well as the possibility of Microsoft adding new properties to the
future versions, let's pin it to an older version which is known for
working.

Refers to SPSTRAT-397