Fix update_skus bug when a single gen is set
This commit fixes a bug on `update_skus` which happens whenever a plan
has just a single generation set on SKU.

In the previous versions it was looking only for the SKU ID suffix to
determine the default and alternate gen, which was enough for all cases
with 2 generations defined but would cause a bug (renaming the SKU ID)
whenever a single generation was originally set.

With this fix it looks not only in the SKU ID suffix but also on SKU
image type to determine which generation is the default and which is the
alternate to prevent renaming the default.

Refers to SPSTRAT-462