Azure: Add support for ARM64 Images
This commit introduces the support for publishing ARM64 images on Azure
Marketplace through the product ingestion (Graph) API.

Since `arm64` images just uses `Gen2` it only preserves the logic of
multiple generations for `x64` and introduces a new SKU naming pattern
of: "{planID}-{arch}" for `arm64`.

With this change, the `publish` method is able to process either `x64`
or `arm64` images, one per call, as the `metadata.architecture` must
indicate one or another when publishing.

This commit also implements several unit tests to ensure the new
funcionlity works as intended and doesn't break the existing one for
`x64`.