Azure: Fix bug on update_skus
This commit fixes the decision criteria to return the existing SKUs
instead of creating new ones on `update_skus`.

The old command was trying to infer the number of the disk images for
each version, which is not the correct way of doing this nor was
properly working.

The new criteria is the following:

1. The private method _all_skus_present will evaluate which image types
   are already declared in the SKUs
2. If there's a disk versino which contains an image which declares a
   type not already present in the SKUs it will return `False`,
   indicating the need to generate the new SKUs and `True` otherwise.

Refers to SPSTRAT-514

Signed-off-by: Jonathan Gangi <jgangi@redhat.com>