Azure: Allow partial SAS comparison
This commit changes the Azure's implementation `utils.is_sas_present`
function to have a new boolean argument named `base_only` which defaults
to `False`. When set as `True` it will cause the SAS comparison to not
check for its arguments equivalence, rather just whether the SAS URI
part is the same or not.

It also changes the `AzurePublishingMetadata` to include a new property
named `check_base_sas_only` to transmit its boolean value to the
`utils.is_sas_present` function.

In practice, this allows the library to have a more permissive SAS
comparison when required while maintaning the existing behavior as
default.

Signed-off-by: Jonathan Gangi <jgangi@redhat.com>