Azure: Adds unit test for publishing offer in preview
This commit adds a new unit test named
`test_publish_live_when_state_is_preview` which aims to ensure that
`cloudpub` is capable of publishing content to Azure when the offer is
in the `preview` state with the desired SAS URI already assossiated with
it.

The expected behavior is to publish live when such events happens.

Refers to SPSTRAT-583

Signed-off-by: Jonathan Gangi <jgangi@redhat.com>