Azure: Allow modular publishing
This commit introduces a new feature for Azure named "modular publish",
which allows the library to only submit the changed technical config for
a given offer/plan to "preview/live" instead of the whole offer like
before.

The new feature is disabled by default in order to keep the library
consistent with its past behavior, being possible to activate it by
setting the value `modular_push = True` on `AzurePublishingMetadata`.

Refers to SPSTRAT-604

Signed-off-by: Jonathan Gangi <jgangi@redhat.com>