Azure: check for SAS in all targets
This commit refactors the `AzureService.publish` to look up for the
existing SAS uri in all targets (`preview`, `live`, `draft`) before
assuming it's not present and making changes.

It keeps the behavior of `overwrite` which always adds the new SAS URI
to the plan, regardless of finding it or not.

Refers to SPSTRAT-595

Signed-off-by: Jonathan Gangi <jgangi@redhat.com>