azure: simplify the submission targets usage
This commit simplifies the methods from `AzureService` to require a
single `target` whenever retrieving a `product` or `plan`, instead of
automatically trying to find it from all targets.

With that the `publish` function will always have to pass the correct
targets and the code gets way more predictable and simplified.

It also helps to reduce the amount of requests for the Microsoft Graph
API as the `get_*` will always perform a single request to retrieve the
data in a particular target.

However, there's a tiny flaw: It will always raise `NotFoundError`
whenever an unexisintg target is passed. Not all products have all 3
targets in their submission states, so this issue will be addresed in
the next commit.

Refers to SPSTRAT-595

Signed-off-by: Jonathan Gangi <jgangi@redhat.com>