Azure: wait for active publishing before executing
THis commit creates a new method for `AzureService` named
`wait_active_publishing` which relies on `ensure_can_publish` to verify
whether changes are being made or not and wait until it's possible to
proceed or timeout if necessary.

The timeout and retry interval are now possible to be set during the
class construction as optional parameters.

With this the `publish` method will double check before doing any
changes:

1. Before loading the product information it will wait for active
   changes so it can retrieve the "latest" content

2. During the publishing phase, if it's no longer the latest change it
   will raise from `ensure_can_publish`

Refers to SPSTRAT-611 and SPSTRAT-549

Signed-off-by: Jonathan Gangi <jgangi@redhat.com>